split trainer_utils into collate and trainer.setup
