+++ tests/test_functional/test_trainer_utils.py
@@ -1,10 +1,10 @@
 import unittest
 from corebehrt.modules.setup.config import Config
-from corebehrt.functional.trainer_utils import (
+from corebehrt.functional.trainer.setup import (
     replace_steps_with_epochs,
     convert_epochs_to_steps,
-    dynamic_padding,
 )
+from corebehrt.functional.collate import dynamic_padding
 import torch
